Add cleanup workflow and update ETL configurations for DuckDB integration
- Introduced a new cleanup workflow to drop intermediate schemas and vacuum the DuckDB database.
- Updated ETL configuration files to set the OMOP schema to 'main' and added cleanup parameters.
- Modified the publish workflow to materialize final OMOP tables instead of creating views.
- Adjusted tests to reflect schema changes and validate the new cleanup functionality.

+++ test/mock_data/generate_mock_data.py
@@ -361,7 +361,7 @@ def build_sources(base_dir: Path) -> None:
 
 def map_type(bq_type: str) -> str:
     return {
-        "INT64": "BIGINT",
+        "INT64": "INTEGER",
         "STRING": "VARCHAR",
         "FLOAT64": "DOUBLE",
         "DATE": "DATE",
